requirements.txt
flask==2.3.3
pdfplumber==0.10.2
plotly==5.18.0
pandas==2.1.2
numpy==1.26.1
jieba==0.42.1
requests==2.31.0
python-dotenv==1.0.0
wordcloud==1.9.2
matplotlib==3.8.1
openai==1.3.7
langdetect==1.0.9
chardet==5.2.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.2.0
beautifulsoup4==4.12.2